% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{download_raw_data}
\alias{download_raw_data}
\title{Download compressed GTEx expression dataset (.gct.gz)}
\usage{
download_raw_data(download_dir = NULL, overwrite = F, gtex_version = "v8")
}
\arguments{
\item{download_dir}{target local directory for download file}

\item{overwrite}{do not overwrite (FALSE) if expression file exist, or overwrite (TRUE)}

\item{gtex_version}{version of gtex database}
}
\description{
Download compressed GTEx expression dataset (.gct.gz)
}
\keyword{internal}
