% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geneXref}
\alias{geneXref}
\title{Cross-reference between gene symbols and Ensembl gene identifiers}
\format{
A data.frame with two columns, and 56,200 rows:
\itemize{
  \item \emph{symbol} - Primary gene symbol
  \item \emph{ensembl_gene_id} - Ensembl gene identifier
}
}
\usage{
geneXref
}
\description{
Cross-reference between gene symbols and Ensembl gene identifiers
}
\keyword{datasets}
