% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_sample_identifiers}
\alias{get_sample_identifiers}
\title{Function that retrieves GTex RNA-seq sample identifiers
for a given tissue type}
\usage{
get_sample_identifiers(
  tissue = "Adrenal Gland",
  analysis = "RNASEQ",
  max_autolysis_score = 3
)
}
\arguments{
\item{tissue}{Type of tissue. Either of the following: 'Blood','Brain',
'Adipose Tissue','Muscle','Blood Vessel','Heart','Ovary','Uterus','Vagina',
'Breast','Skin','Salivary Gland','Adrenal Gland','Thyroid','Lung','Spleen',
'Pancreas','Espohagus','Stomach','Colon','Small Intestine','Prostate',
'Testis','Nerve','Pituitary','Liver','Kidney','Cervix Uteri','Fallopian Tube',
'Bladder','Bone Marrow'}

\item{analysis}{RNASEQ}

\item{max_autolysis_score}{Set maximum allowed autolysis score for samples
retrieved (1 (None) to 3 (Severe))}
}
\description{
Function that retrieves GTex RNA-seq sample identifiers
for a given tissue type
}
\keyword{internal}
