% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_tpm_data}
\alias{get_tpm_data}
\title{Retrieve sample gene expression data from GTEx}
\usage{
get_tpm_data(
  tissue_type = "Breast_Mammary_Tissue",
  genes = NULL,
  gtex_version = "v8"
)
}
\arguments{
\item{tissue_type}{tissue type}

\item{genes}{character vector with primary gene symbols}

\item{gtex_version}{version of GTEx}
}
\description{
Retrieves sample gene expression data from GTEx, using the public API.
A tissue type must be provided, and also a character vector with genes of interest.
Gene expression estimates are provided for all samples of that tissue type,
with two different expression units provided (TPM, log2(TPM + 1))
}
