% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_tpm_data_api}
\alias{get_tpm_data_api}
\title{Function that retrieves gene expression data for a given set of
genes from GTEx, using the public API. A tissue type must be provided
Gene expression estimates are provided for all samples of that tissue type,
with two different expression units provided (TPM, log2(TPM + 1))}
\usage{
get_tpm_data_api(
  tissue_type = "Breast_Mammary_Tissue",
  genes = NULL,
  gtex_version = "v8"
)
}
\arguments{
\item{tissue_type}{tissue type}

\item{genes}{character vector with primary gene symbols@}

\item{gtex_version}{version of GTEx}
}
\description{
Function that retrieves gene expression data for a given set of
genes from GTEx, using the public API. A tissue type must be provided
Gene expression estimates are provided for all samples of that tissue type,
with two different expression units provided (TPM, log2(TPM + 1))
}
