% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleMetadata}
\alias{sampleMetadata}
\title{Sample properties of GTEx RNAseq samples (v8)}
\format{
A data.frame with seven columns, and 17,832 rows:
\itemize{
  \item \emph{sample_id} - GTEx sample identifier
  \item \emph{tissue_type} - Tissue type annotation
  \item \emph{tissue_type_detailed} - Detailed annotation of tissue type
  \item \emph{in_analysis_freeze} - Type of analysis (RNASEQ)
  \item \emph{autolysis_score} - Autolysis score (1 to 3)
  \item \emph{tissue_site_detail_id} - Slight modification of 'tissue_type_detailed', used for querying API
  \item \emph{num_tissue_samples} - Number of samples for a given 'tissue_site_detail_id'
  \item \emph{tcga_study_match} - Matching TCGA study (for type of tissue)
}
}
\usage{
sampleMetadata
}
\description{
Sample properties of GTEx RNAseq samples (v8)
}
\keyword{datasets}
